Only allow to update or delete own posts

+++ src/routers/posts.py
@@ -52,9 +52,12 @@ def create_post(
         current_user: User = Depends(get_current_user)
 ):
     # new_post = Post(title=post.title, content=post.content, published=post.published)
-    print(current_user.email)
+
+    # Create a new post with data from the request body
     new_post = Post(**post.model_dump())
-    
+    # Add an owner_id from the currently logged-in user
+    new_post.owner_id = current_user.id
+
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -64,10 +67,10 @@ def create_post(
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
 def delete_post(
-        id: int, db: Session = Depends(get_db),
+        id: int,
+        db: Session = Depends(get_db),
         current_user: User = Depends(get_current_user)
 ):
-    print(current_user.email)
     
     #? ORM-Style deletion
 
@@ -80,16 +83,24 @@ def delete_post(
     if not post_to_delete:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                         detail=f"Can't find a post with ID {id}")
-    db.delete(post_to_delete)
 
-    #? More efficient, but without ORM Events triggered. Pretty sure, this way is used for SQLAlchemy 1.x
-    # post_to_delete = db.get(Post, id)
-    # if not post_query.first():
-    #     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-    #                     detail=f"Can't find a post with ID {id}")
-    # post_query.delete(synchronize_session=False)
-    
+    if current_user.id != post_to_delete.owner_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f"Not authorized to perform the request")
+
+    db.delete(post_to_delete)
     db.commit()
+
+
+        # ? More efficient, but without ORM Events triggered. Pretty sure, this way is used for SQLAlchemy 1.x
+        # post_to_delete = db.get(Post, id)
+        # if not post_query.first():
+        #     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+        #                     detail=f"Can't find a post with ID {id}")
+        # post_query.delete(synchronize_session=False)
+        # db.commit()
+    
+
     return
 
 @router.put("/{id}", status_code=status.HTTP_201_CREATED, response_model=PostResponse)
@@ -101,17 +112,29 @@ def update_post(
     print(current_user.email)
     #? Non-ORM Update
 
-    #? Explicit way of writing the query
-    stmt = (update(Post)
-            .where(Post.id == id)
-            .values(**post.model_dump())
-            .returning(Post)
-            )
-            
-    post_to_update = db.execute(stmt).scalar_one_or_none()
+    post_to_update = db.get(Post, id)
+
     if not post_to_update:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                         detail=f"Can't find a post with ID {id}")
-    
+
+    if current_user.id != post_to_update.owner_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f"Not authorized to perform the request")
+
+    #? ORM way of updating the values
+    for field, value in post.model_dump().items():
+        setattr(post_to_update, field, value)
+
+    #? Explicit way of writing the query
+    # stmt = (update(Post)
+    #         .where(Post.id == id)
+    #         .values(**post.model_dump())
+    #         .returning(Post)
+    #         )
+    #
+    # post_to_update = db.execute(stmt).scalar_one_or_none()
+
     db.commit()
+    db.refresh(post_to_update)
     return post_to_update
